@@ -29,6 +29,7 @@ import android.util.SparseLongArray;
 
 import com.telink.ble.mesh.core.Encipher;
 import com.telink.ble.mesh.core.MeshUtils;
+import com.telink.ble.mesh.core.ble.GattConnection;
 import com.telink.ble.mesh.core.message.MeshMessage;
 import com.telink.ble.mesh.core.message.Opcode;
 import com.telink.ble.mesh.core.networking.beacon.MeshBeaconPDU;
@@ -86,7 +87,7 @@ public class NetworkingController {
 
 //    private static final int SEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH = UNSEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH + 1;
 
-    private ExtendBearerMode extendBearerMode;
+    private ExtendBearerMode extendBearerMode = ExtendBearerMode.NONE;
 
     // segmentedAccessLength = unsegmentedAccessLength + 1
 //    public int unsegmentedAccessLength = UNSEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH_DEFAULT;
@@ -151,7 +152,8 @@ public class NetworkingController {
     /**
      * from mesh configuration
      */
-    private int initIvIndex = 0;
+    private long initIvIndex = 0;
+
 
     /**
      * unsigned 32-bit integer
@@ -246,6 +248,8 @@ public class NetworkingController {
 
     private Set<Integer> mResponseMessageBuffer = new LinkedHashSet<>();
 
+    private int[] whiteList;
+
     private int proxyFilterInitStep = 0;
 
     private static final int PROXY_FILTER_INIT_STEP_SET_TYPE = 1;
@@ -265,6 +269,8 @@ public class NetworkingController {
     public static final long NETWORK_INTERVAL_FOR_FU = 180; // 240 ms // 320
 
     public static final long NETWORK_INTERVAL_DEFAULT = 240; // 240 ms // 320
+
+//    public static final long NETWORK_INTERVAL_DEFAULT = 10; // for test
 
     /**
      * network packet sent to un-direct connected node should push to queue, and send periodically
@@ -291,8 +297,8 @@ public class NetworkingController {
     public void setup(MeshConfiguration configuration) {
         this.clear();
         this.resetDirectAddress();
-        this.initIvIndex = configuration.ivIndex;
-        this.ivIndex = initIvIndex & MeshUtils.UNSIGNED_INTEGER_MAX;
+        this.initIvIndex = configuration.ivIndex & MeshUtils.UNSIGNED_INTEGER_MAX;
+        this.ivIndex = initIvIndex;
         int seqNo = configuration.sequenceNumber;
         this.mSequenceNumber.set(initSequenceNumber(seqNo));
         this.netKeyIndex = configuration.netKeyIndex;
@@ -305,6 +311,7 @@ public class NetworkingController {
         this.deviceKeyMap = configuration.deviceKeyMap;
 
         this.localAddress = configuration.localAddress;
+        this.whiteList = configuration.proxyFilterWhiteList;
     }
 
 
@@ -347,6 +354,9 @@ public class NetworkingController {
     }
 
     private int getSegmentAccessLength(int dstAddress, int opcode) {
+        if (GattConnection.mtu < UNSEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH_LONG) {
+            return UNSEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH_DEFAULT;
+        }
         if (dstAddress == directAddress && opcode == Opcode.BLOB_CHUNK_TRANSFER.value) {
             return UNSEGMENTED_ACCESS_PAYLOAD_MAX_LENGTH_LONG;
         }
@@ -405,10 +415,10 @@ public class NetworkingController {
         }
 
         if (privateBeaconReceived) {
-            MeshPrivateBeacon beacon = MeshPrivateBeacon.createIvUpdatingBeacon(this.initIvIndex, privateBeaconKey, isIvUpdating);
+            MeshPrivateBeacon beacon = MeshPrivateBeacon.createIvUpdatingBeacon((int) this.initIvIndex, privateBeaconKey, isIvUpdating);
             sendMeshBeaconPdu(beacon);
         } else {
-            SecureNetworkBeacon networkBeacon = SecureNetworkBeacon.createIvUpdatingBeacon(this.initIvIndex, networkId, beaconKey, isIvUpdating);
+            SecureNetworkBeacon networkBeacon = SecureNetworkBeacon.createIvUpdatingBeacon((int) this.initIvIndex, networkId, beaconKey, isIvUpdating);
             log("send beacon: " + networkBeacon.toString());
             sendMeshBeaconPdu(networkBeacon);
         }
@@ -417,7 +427,7 @@ public class NetworkingController {
     }
 
     private void onIvUpdated(long newIvIndex) {
-        if (newIvIndex > initIvIndex) {
+        if (newIvIndex > initIvIndex || this.initIvIndex == MeshUtils.IV_MISSING) {
             log(String.format(" iv updated to %08X", newIvIndex));
             this.initIvIndex = (int) newIvIndex;
             this.deviceSequenceNumberMap.clear();
@@ -436,6 +446,12 @@ public class NetworkingController {
                 updating,
                 this.ivIndex,
                 this.isIvUpdating));
+        if (this.ivIndex == MeshUtils.IV_MISSING) {
+            this.isIvUpdating = updating;
+            this.ivIndex = remoteIvIndex;
+            this.onIvUpdated(remoteIvIndex);
+            return;
+        }
         // d-value
         long dVal = remoteIvIndex - this.ivIndex;
 
@@ -996,7 +1012,11 @@ public class NetworkingController {
                 this.directAddress = src;
                 proxyFilterInitStep++;
                 if (proxyFilterInitStep == PROXY_FILTER_INIT_STEP_SET_TYPE) {
-                    addFilterAddress(new int[]{localAddress, 0xFFFF});
+                    if (this.whiteList == null || this.whiteList.length == 0) {
+                        addFilterAddress(new int[]{localAddress, MeshUtils.ADDRESS_BROADCAST});
+                    } else {
+                        addFilterAddress(this.whiteList);
+                    }
                 } else if (proxyFilterInitStep == PROXY_FILTER_SET_STEP_ADD_ADR) {
                     onProxyInitComplete(true);
                 }
